+++ app/features/dynamo/prompt/summarize-structured-tabular-data-prompt.txt
