--- app/features/dynamoExtendedFileSupport/metadata.json
@@ -1,14 +1,0 @@
-{
-    "inputs": [
-        {
-            "label": "File Url",
-            "name": "file_url",
-            "type": "string"
-        },
-        {
-            "label": "File Type",
-            "name": "file_type",
-            "type": "string"
-        }
-    ]
-}
