--- app/features/dynamoExtendedFileSupport/prompt/examples.txt
@@ -1,60 +1,0 @@
-input:
-## Concise Summary of the provided document:
-
-**Large Language Models (LLMs)** are powerful AI tools trained on massive datasets to perform tasks like text generation, translation, and question answering. They can be specialized for specific domains through fine-tuning, making them versatile and adaptable. 
-
-**Key points:**
-
-* **Pre-trained and fine-tuned:** LLMs learn general knowledge from large datasets and specialize in specific tasks through additional training. 
-* **Prompt design:** Effective prompts are crucial for eliciting desired responses from LLMs.
-* **Domain knowledge:** Understanding the specific domain is essential for building and tuning LLMs.
-* **Parameter-efficient tuning methods:** This method allows for efficient customization of LLMs without altering the entire model.
-* **Vertex AI:** Provides tools for building, tuning, and deploying LLMs for specific tasks.
-* **Generative AI App Builder and PaLM API:** Tools for developers to build AI apps and experiment with LLMs.
-* **Model management tools:** Tools for training, deploying, and monitoring ML models.
-
-**This document provides a comprehensive overview of LLMs and related tools, highlighting their capabilities and potential applications.** 
-
-**Additional notes:**
-
-* The text emphasizes the importance of prompt design and domain knowledge for effective LLM usage.
-* It introduces cutting-edge technologies like PETM and Vertex AI, showcasing the rapid advancements in the field.
-* The document also provides practical resources for developers to build and deploy LLM-powered applications. 
-
-**Overall, this document is a valuable resource for anyone interested in understanding and utilizing LLMs and related technologies.**
-
-Output:
-[
-  {
-    "concept": "Large Language Models (LLMs)",
-    "definition": "Powerful AI tools trained on massive datasets to perform tasks like text generation, translation, and question answering."
-  },
-  {
-    "concept": "Pre-trained and fine-tuned",
-    "definition": "LLMs learn general knowledge from large datasets and specialize in specific tasks through additional training."
-  },
-  {
-    "concept": "Prompt design",
-    "definition": "Effective prompts are crucial for eliciting desired responses from LLMs."
-  },
-  {
-    "concept": "Domain knowledge",
-    "definition": "Understanding the specific domain is essential for building and tuning LLMs."
-  },
-  {
-    "concept": "Parameter-efficient tuning methods",
-    "definition": "This method allows for efficient customization of LLMs without altering the entire model."
-  },
-  {
-    "concept": "Vertex AI",
-    "definition": "Provides tools for building, tuning, and deploying LLMs for specific tasks."
-  },
-  {
-    "concept": "Generative AI App Builder and PaLM API",
-    "definition": "Tools for developers to build AI apps and experiment with LLMs."
-  },
-  {
-    "concept": "Model management tools",
-    "definition": "Tools for training, deploying, and monitoring ML models."
-  }
-]
